+++ src/windows/player.c
@@ -258,7 +258,7 @@ static void window_player_overview_show_group_dropdown(rct_window *w, rct_widget
 
 	for (i = 0; i < network_get_num_groups(); i++) {
 		gDropdownItemsFormat[i] = STR_OPTIONS_DROPDOWN_ITEM;
-		gDropdownItemsArgs[i] = (int)network_get_group_name(i);
+		gDropdownItemsArgs[i] = (uintptr_t)network_get_group_name(i);
 	}
 
 	dropdown_set_checked(network_get_group_index(network_get_player_group(player)), true);
